political_perspective_app/requirements.txt
# Web Framework & UI
streamlit==1.53.0
fastapi==0.135.1
uvicorn==0.41.0
requests==2.32.5

# LLM & LangChain Ecosystem
langchain==1.2.12
langchain-core==1.2.19
langchain-openai==1.1.11
langchain-anthropic==1.4.0
langchain-google-genai==4.2.1
langgraph==1.1.2
openai==2.26.0

# Database & Data Analysis
SQLAlchemy==2.0.45
pandas==2.3.3

# Environment, Utilities & Monitoring
pydantic==2.12.5
python-dotenv==1.2.1
langsmith==0.7.20